requirements.txt
django==2.2.5
pep8==1.7.1
flake8==3.7.8
django-widget-tweaks==1.4.1
django-factory-boy==1.0.0
pillow==6.1.0
sorl-thumbnail==12.5.0
python-magic==0.4.10
django-cleanup==0.4.0
pycrypto==2.6.1
freezegun==0.3.5
django-inplaceedit==1.4.1
django-debug-toolbar==2.0
django-haystack==2.4.1
Whoosh==2.7.4
django-inplaceedit-bootstrap==0.2.1
django-extensions==1.6.1
werkzeug==0.16.0
pytube==9.5.2
requests==2.22.0
structlog==0.4.1
celery==3.1.19
raven==5.9.2
psycopg2==2.8.3
librabbitmq==2.0.0
